Move SpecFilter to `spack.solver.reuse`
