+++ lib/spack/spack/solver/core.py
@@ -1,12 +1,13 @@
 # Copyright Spack Project Developers. See COPYRIGHT file for details.
 #
 # SPDX-License-Identifier: (Apache-2.0 OR MIT)
-"""Low-level wrappers around clingo API."""
+"""Low-level wrappers around clingo API and other basic functionality related to ASP"""
 import importlib
 import pathlib
 from types import ModuleType
 from typing import Any, Callable, NamedTuple, Optional, Tuple, Union
 
+import spack.platforms
 from spack.llnl.util import lang
 
 
@@ -299,3 +300,8 @@ class SourceContext:
         # in that case).
         self.source = "none" if source is None else source
         self.wrap_node_requirement: Optional[bool] = None
+
+
+def using_libc_compatibility() -> bool:
+    """Returns True if we are currently using libc compatibility"""
+    return spack.platforms.host().name == "linux"
